# API_Airbnb/API_Airbnb/bin/Release/net6.0/swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "API_Airbnb, Version=1.0.0.0, Culture=neutral, PublicKeyToken=null",
    "version": "1.0"
  },
  "paths": {
...
    "/api/ArProperties/Delete/{Id}": {
      "put": {
        "tags": [
          "ArProperties"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "Id",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/ArProperties/{Id}": {
      "get": {
        "tags": [
          "ArProperties"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "Id",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/ArProperties": {
      "post": {
        "tags": [
          "ArProperties"
        ],
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/PropertyDTO"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/ArProperties/Update/{Id}": {
      "put": {
        "tags": [
          "ArProperties"
        ],
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "Id",
            "required": true,
            "type": "integer",
            "format": "int32"
          },
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/PropertyDTO"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
...
    "/api/ArProperties/GetWithAllData/{Id}": {
      "get": {
        "tags": [
          "ArProperties"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "Id",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
...
  },
  "definitions": {
...
    "PropertyDTO": {
      "required": [
        "accomodatesCount",
        "address",
        "bathroomCount",
        "bedCount",
        "bedroomCount",
        "categoryId",
        "cityId",
        "countryId",
        "currencyId",
        "description",
        "name",
        "price",
        "propertyTypeId",
        "roomTypeId"
      ],
      "type": "object",
      "properties": {
        "name": {
          "minLength": 1,
          "type": "string"
        },
        "description": {
          "minLength": 1,
          "type": "string"
        },
        "propertyTypeId": {
          "format": "int32",
          "type": "integer"
        },
        "roomTypeId": {
          "format": "int32",
          "type": "integer"
        },
        "categoryId": {
          "format": "int32",
          "type": "integer"
        },
        "countryId": {
          "format": "int32",
          "type": "integer"
        },
        "cityId": {
          "format": "int32",
          "type": "integer"
        },
        "address": {
          "minLength": 1,
          "type": "string"
        },
        "latitude": {
          "type": "string"
        },
        "longitude": {
          "type": "string"
        },
        "bedroomCount": {
          "format": "int32",
          "type": "integer"
        },
        "bedCount": {
          "format": "int32",
          "type": "integer"
        },
        "bathroomCount": {
          "format": "int32",
          "type": "integer"
        },
        "accomodatesCount": {
          "format": "int32",
          "type": "integer"
        },
        "availabilityType": {
          "type": "boolean"
        },
        "price": {
          "format": "double",
          "type": "number"
        },
        "currencyId": {
          "format": "int32",
          "type": "integer"
        },
        "priceType": {
          "format": "int32",
          "type": "integer"
        },
        "minimumStay": {
          "type": "string"
        },
        "minimumStayType": {
          "format": "int32",
          "type": "integer"
        },
        "refundType": {
          "type": "boolean"
        },
        "url": {
          "type": "string"
        }
      }
    },
...
  }
}